Minor changes to stored data types Added population procedure from directory of CSV files

+++ DB/DDL/f1/tables/constructor_results.sql
@@ -3,6 +3,6 @@ create table f1.constructor_results (
     id int,
     race_id int,
     constructor_id int,
-    points int,
+    points float,
     status char(2)
 );
