+++ DB/DDL/f1/tables/constructor_standings.sql
@@ -3,7 +3,7 @@ create table f1.constructor_standings (
     id int,
     race_id int,
     constructor_id int,
-    points int,
+    points float,
     position int,
     position_text varchar(5),
     wins int
